requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
discord==2.3.2
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
Pygments==2.19.2
python-env==1.0.0
requests==2.32.5
rich==14.2.0
soupsieve==2.8.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
Wikipedia-API==0.8.1
yarl==1.22.0
